Moving to conf system - CoordinateGridMixin 11
