+++ src/base/i-coordinate-grid-mixin-conf.ts
@@ -4,6 +4,7 @@ import {RoundFn, Units} from '../core/types';
 import {CountableTimeInterval} from 'd3-time';
 
 export interface ICoordinateGridMixinConf extends IMarginMixinConf, IColorMixinConf {
+    readonly useRightYAxis?: boolean;
     readonly clipPadding?: number;
     readonly mouseZoomable?: boolean;
     readonly zoomOutRestrict?: boolean;
